apps/agents/pyproject.toml
[project]
name = "travel-sync-agents"
version = "0.1.0"
description = "Travel Sync AI agents"
requires-python = ">=3.11"
dependencies = [
  "exa-py",
  "firecrawl-py",
  "langchain",
  "langchain-openai",
  "pydantic",
  "python-dotenv",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/travel_sync_agents"]

